Reworked from ground up
I'm away for the next 3 weeks, so I was not able to finish everything. Some flavour from CMP is not included atm, but every mod included was updated (not all of the state traits are added yet, the files are there, but they need to be added to the map_data files). Early Merry Christmas!